fix read-only gifs and math directories
An installed copy of stex might have all write permission bits removed,
even for the owner of the installed files: this is the case with the
Guix package of stex, for example. If such an installation is used to
initialize the gifs or math directories (e.g. for an out-of-source
build), tar likewise creates them without write permissions, preventing
subdirectories from being created later. Set the user write bit
explicitly to avoid this problem.
Furthermore, tar by default would copy symbolic links in the installed
copy as links, causing chmod to fail or, if permitted, writes to an
improper location. Avoid this using the --dereference option, which is
portable to at least GNU and FreeBSD tar (but only in long form: GNU
has -h, but FreeBSD has -L).
  Mf-stex

Related to https://github.com/racket/racket/pull/4203
Closes https://github.com/dybvig/stex/pull/6